bug: Resolved circular dependencies
